refactor: extract OptimizationJobStoreMixin from PlaybookMixin (sqlite + ABC)
Move the 8 plain-CRUD playbook optimization job/candidate/evaluation methods
out of the shrinking PlaybookMixin into a new OptimizationJobStoreMixin in the
existing playbook/ subpackage, for both the SQLite backend and the storage_base
ABC. Behavior-preserving verbatim move; the mixin is composed into SQLiteStorage
and BaseStorage immediately before PlaybookMixin, so all method resolution is
unchanged. The bucket-only row converters move alongside the methods; shared
helpers stay in _playbook.py and are imported.
